Added text to fits conversion code
Added the python script that is required for preparing the observations for GalSpecFitX. It is used for converting txt files with wavelength, flux and error columns into fits files. You can input a folder full of such txt observations files and the code will produce a folder full of fits files ready for the run.